playlist for august 7 2014
fixed utf encoding
also this playlist is the bomb. special thanks to daniel haaser for the KILN recommendation
i also want to give a shoutout to starsmith because i'm OBSESSED with his remixes
and also you'll notice that vaults has been on this for a while. major props to them